refactor: 입력된 핀수를 심볼로 변경
move class
