+++ src/main/java/bowling/Pitching.java
@@ -1,0 +1,25 @@
+package bowling;
+
+public class Pitching extends State {
+  private int first;
+
+  public Pitching(int first) {
+    super();
+    this.first = first;
+  }
+
+  public State roll(int second) {
+    if (first + second ==10) {
+      return new Spare(first, second);
+    }
+    return new Open(first, second);
+  }
+
+  public String symbol() {
+    if (first == 0) {
+      return "-";
+    }
+    return String.valueOf(first);
+  }
+
+}
