+++ src/main/java/bowling/Strike.java
@@ -1,0 +1,9 @@
+package bowling;
+
+public class Strike extends State {
+
+  public String symbol() {
+    return "X";
+  }
+
+}
