+++ src/test/java/bowling/StateTest.java
@@ -63,83 +63,4 @@ public class StateTest {
     assertThat(open.symbol()).isEqualTo("8|-");
   }
 
-  private static class State {
-
-    public String symbol() {
-      return "";
-    }
-
-    public static State of(int pins) {
-      if (pins == 10) {
-        return new Strike();
-      }
-
-      return new Pitching(pins);
-    }
-
-    public State roll(int second) {
-      return null;
-    }
-  }
-
-  private static class Strike extends State {
-
-    public String symbol() {
-      return "X";
-    }
-  }
-
-  private static class Pitching extends State {
-    private int first;
-
-    public Pitching(int first) {
-      super();
-      this.first = first;
-    }
-
-    public State roll(int second) {
-      if (first + second ==10) {
-        return new Spare(first, second);
-      }
-      return new Open(first, second);
-    }
-
-    public String symbol() {
-      if (first == 0) {
-        return "-";
-      }
-      return String.valueOf(first);
-    }
-
-  }
-
-  private static class Spare extends Pitching {
-
-    private int second;
-
-    public Spare(int first, int second) {
-      super(first);
-      this.second = second;
-    }
-
-    public String symbol() {
-      return super.symbol() + "|/";
-    }
-  }
-
-  private static class Open extends Pitching {
-    private int second;
-
-    public Open(int first, int second) {
-      super(first);
-      this.second = second;
-    }
-
-    public String symbol() {
-      if (second == 0) {
-        return super.symbol() + "|-";
-      }
-      return super.symbol() + "|" + second;
-    }
-  }
 }
